feat: added oauth2.0 support

+++ client/src/test/java/com/tigrisdata/db/client/grpc/AuthEnabledTestTigrisService.java
@@ -14,61 +14,54 @@
 package com.tigrisdata.db.client.grpc;
 
 import com.tigrisdata.db.api.v1.grpc.Api;
+import io.grpc.Status;
+import io.grpc.StatusRuntimeException;
 import io.grpc.stub.StreamObserver;
 
-import java.util.UUID;
-
-public class TransactionTestUserService extends TestUserService {
-
-  private String txId;
-  private String txOrigin;
+public class AuthEnabledTestTigrisService extends TestTigrisService {
+  public AuthEnabledTestTigrisService() {
+    super();
+  }
 
   @Override
   public void beginTransaction(
       Api.BeginTransactionRequest request,
       StreamObserver<Api.BeginTransactionResponse> responseObserver) {
-    txId = UUID.randomUUID().toString();
-    txOrigin = txId + "_origin";
-    responseObserver.onNext(
-        Api.BeginTransactionResponse.newBuilder()
-            .setTxCtx(Api.TransactionCtx.newBuilder().setId(txId).setOrigin(txOrigin).build())
-            .build());
-    responseObserver.onCompleted();
+    if (!containsAuthHeaders()) {
+      responseObserver.onError(new StatusRuntimeException(Status.UNAUTHENTICATED));
+      return;
+    }
+    super.beginTransaction(request, responseObserver);
   }
 
   @Override
   public void commitTransaction(
       Api.CommitTransactionRequest request,
       StreamObserver<Api.CommitTransactionResponse> responseObserver) {
-    if (isValidTransactionState()) {
-      responseObserver.onError(new IllegalStateException("Transaction is not active"));
+    if (!containsAuthHeaders()) {
+      responseObserver.onError(new StatusRuntimeException(Status.UNAUTHENTICATED));
       return;
     }
-    resetTx();
-    txId = UUID.randomUUID().toString();
-    txOrigin = txId + "_origin";
-    responseObserver.onNext(Api.CommitTransactionResponse.newBuilder().build());
-    responseObserver.onCompleted();
+    super.commitTransaction(request, responseObserver);
   }
 
   @Override
   public void rollbackTransaction(
       Api.RollbackTransactionRequest request,
       StreamObserver<Api.RollbackTransactionResponse> responseObserver) {
-    if (isValidTransactionState()) {
-      responseObserver.onError(new IllegalStateException("Transaction is not active"));
+    if (!containsAuthHeaders()) {
+      responseObserver.onError(new StatusRuntimeException(Status.UNAUTHENTICATED));
       return;
     }
-    resetTx();
-    responseObserver.onNext(Api.RollbackTransactionResponse.newBuilder().build());
-    responseObserver.onCompleted();
+    super.rollbackTransaction(request, responseObserver);
   }
 
   @Override
   public void insert(
       Api.InsertRequest request, StreamObserver<Api.InsertResponse> responseObserver) {
-    if (isValidTransactionState()) {
-      responseObserver.onError(new IllegalStateException("Transaction is not active"));
+    if (!containsAuthHeaders()) {
+      responseObserver.onError(new StatusRuntimeException(Status.UNAUTHENTICATED));
+      return;
     }
     super.insert(request, responseObserver);
   }
@@ -76,8 +69,9 @@ public class TransactionTestUserService extends TestUserService {
   @Override
   public void delete(
       Api.DeleteRequest request, StreamObserver<Api.DeleteResponse> responseObserver) {
-    if (isValidTransactionState()) {
-      responseObserver.onError(new IllegalStateException("Transaction is not active"));
+    if (!containsAuthHeaders()) {
+      responseObserver.onError(new StatusRuntimeException(Status.UNAUTHENTICATED));
+      return;
     }
     super.delete(request, responseObserver);
   }
@@ -85,8 +79,9 @@ public class TransactionTestUserService extends TestUserService {
   @Override
   public void update(
       Api.UpdateRequest request, StreamObserver<Api.UpdateResponse> responseObserver) {
-    if (isValidTransactionState()) {
-      responseObserver.onError(new IllegalStateException("Transaction is not active"));
+    if (!containsAuthHeaders()) {
+      responseObserver.onError(new StatusRuntimeException(Status.UNAUTHENTICATED));
+      return;
     }
     super.update(request, responseObserver);
   }
@@ -94,16 +89,18 @@ public class TransactionTestUserService extends TestUserService {
   @Override
   public void replace(
       Api.ReplaceRequest request, StreamObserver<Api.ReplaceResponse> responseObserver) {
-    if (isValidTransactionState()) {
-      responseObserver.onError(new IllegalStateException("Transaction is not active"));
+    if (!containsAuthHeaders()) {
+      responseObserver.onError(new StatusRuntimeException(Status.UNAUTHENTICATED));
+      return;
     }
     super.replace(request, responseObserver);
   }
 
   @Override
   public void read(Api.ReadRequest request, StreamObserver<Api.ReadResponse> responseObserver) {
-    if (isValidTransactionState()) {
-      responseObserver.onError(new IllegalStateException("Transaction is not active"));
+    if (!containsAuthHeaders()) {
+      responseObserver.onError(new StatusRuntimeException(Status.UNAUTHENTICATED));
+      return;
     }
     super.read(request, responseObserver);
   }
@@ -112,8 +109,9 @@ public class TransactionTestUserService extends TestUserService {
   public void createOrUpdateCollection(
       Api.CreateOrUpdateCollectionRequest request,
       StreamObserver<Api.CreateOrUpdateCollectionResponse> responseObserver) {
-    if (isValidTransactionState()) {
-      responseObserver.onError(new IllegalStateException("Transaction is not active"));
+    if (!containsAuthHeaders()) {
+      responseObserver.onError(new StatusRuntimeException(Status.UNAUTHENTICATED));
+      return;
     }
     super.createOrUpdateCollection(request, responseObserver);
   }
@@ -122,8 +120,9 @@ public class TransactionTestUserService extends TestUserService {
   public void dropCollection(
       Api.DropCollectionRequest request,
       StreamObserver<Api.DropCollectionResponse> responseObserver) {
-    if (isValidTransactionState()) {
-      responseObserver.onError(new IllegalStateException("Transaction is not active"));
+    if (!containsAuthHeaders()) {
+      responseObserver.onError(new StatusRuntimeException(Status.UNAUTHENTICATED));
+      return;
     }
     super.dropCollection(request, responseObserver);
   }
@@ -132,8 +131,9 @@ public class TransactionTestUserService extends TestUserService {
   public void listDatabases(
       Api.ListDatabasesRequest request,
       StreamObserver<Api.ListDatabasesResponse> responseObserver) {
-    if (isValidTransactionState()) {
-      responseObserver.onError(new IllegalStateException("Transaction is not active"));
+    if (!containsAuthHeaders()) {
+      responseObserver.onError(new StatusRuntimeException(Status.UNAUTHENTICATED));
+      return;
     }
     super.listDatabases(request, responseObserver);
   }
@@ -142,8 +142,9 @@ public class TransactionTestUserService extends TestUserService {
   public void listCollections(
       Api.ListCollectionsRequest request,
       StreamObserver<Api.ListCollectionsResponse> responseObserver) {
-    if (isValidTransactionState()) {
-      responseObserver.onError(new IllegalStateException("Transaction is not active"));
+    if (!containsAuthHeaders()) {
+      responseObserver.onError(new StatusRuntimeException(Status.UNAUTHENTICATED));
+      return;
     }
     super.listCollections(request, responseObserver);
   }
@@ -152,8 +153,9 @@ public class TransactionTestUserService extends TestUserService {
   public void createDatabase(
       Api.CreateDatabaseRequest request,
       StreamObserver<Api.CreateDatabaseResponse> responseObserver) {
-    if (isValidTransactionState()) {
-      responseObserver.onError(new IllegalStateException("Transaction is not active"));
+    if (!containsAuthHeaders()) {
+      responseObserver.onError(new StatusRuntimeException(Status.UNAUTHENTICATED));
+      return;
     }
     super.createDatabase(request, responseObserver);
   }
@@ -161,26 +163,18 @@ public class TransactionTestUserService extends TestUserService {
   @Override
   public void dropDatabase(
       Api.DropDatabaseRequest request, StreamObserver<Api.DropDatabaseResponse> responseObserver) {
-    if (isValidTransactionState()) {
-      responseObserver.onError(new IllegalStateException("Transaction is not active"));
+    if (!containsAuthHeaders()) {
+      responseObserver.onError(new StatusRuntimeException(Status.UNAUTHENTICATED));
+      return;
     }
     super.dropDatabase(request, responseObserver);
   }
 
-  @Override
-  public void reset() {
-    super.reset();
-    resetTx();
-  }
-
-  private void resetTx() {
-    txId = "";
-    txOrigin = "";
-  }
-
-  private boolean isValidTransactionState() {
-    String incomingTxId = ContextSettingServerInterceptor.TX_ID_CONTEXT_KEY.get();
-    String incomingTxOrigin = ContextSettingServerInterceptor.TX_ORIGIN_CONTEXT_KEY.get();
-    return !txId.equals(incomingTxId) || !txOrigin.equals(incomingTxOrigin);
+  private boolean containsAuthHeaders() {
+    String incomingAuthorizationHeader =
+        ContextSettingServerInterceptor.AUTHORIZATION_CONTEXT_KEY.get();
+    return incomingAuthorizationHeader != null
+        && incomingAuthorizationHeader.equals(
+            "bearer <header>.ewogIAogICJleHAiOiAyOTk5OTk5OTk5Cn0=.<signature>");
   }
 }
